# NameValueCollection.ToQueryString performance optimisation