--- a/src/Elasticsearch.Net/Extensions/NameValueCollectionExtensions.cs
+++ b/src/Elasticsearch.Net/Extensions/NameValueCollectionExtensions.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information
 
 using System;
+using System.Buffers;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Linq;
@@ -12,28 +13,56 @@ namespace Elasticsearch.Net.Extensions
 {
 	internal static class NameValueCollectionExtensions
 	{
+		private const int MaxCharsOnStack = 256; // 512 bytes
+
 		internal static string ToQueryString(this NameValueCollection nv)
 		{
 			if (nv == null || nv.AllKeys.Length == 0) return string.Empty;
 
-			// initialize with capacity for number of key/values with length 5 each
-			var builder = new StringBuilder("?", nv.AllKeys.Length * 2 * 5);
-			for (var i = 0; i < nv.AllKeys.Length; i++)
+			var maxLength = 1 + nv.AllKeys.Length - 1; // account for '?', and any required '&' chars
+			foreach (var key in nv.AllKeys)
 			{
-				if (i != 0)
-					builder.Append("&");
-
-				var key = nv.AllKeys[i];
-				builder.Append(Uri.EscapeDataString(key));
-				var value = nv[key];
-				if (!value.IsNullOrEmpty())
-				{
-					builder.Append("=");
-					builder.Append(Uri.EscapeDataString(nv[key]));
-				}
+				maxLength += 1 + (key.Length + nv[key]?.Length ?? 0) * 3; // '=' char + worst case assume all key/value chars are escaped
 			}
 
-			return builder.ToString();
+			// prefer stack allocated array for short lengths
+			// note: renting for larger lengths is slightly more efficient since no zeroing occurs
+			char[] rentedFromPool = null;
+			var buffer = maxLength > MaxCharsOnStack
+				? rentedFromPool = ArrayPool<char>.Shared.Rent(maxLength)
+				: stackalloc char[MaxCharsOnStack];
+
+			try
+			{
+				var position = 0;
+				buffer[position++] = '?';
+
+				foreach (var key in nv.AllKeys)
+				{
+					if (position != 1)
+						buffer[position++] = '&';
+
+					var escapedKey = Uri.EscapeDataString(key);
+					escapedKey.AsSpan().CopyTo(buffer.Slice(position));
+					position += escapedKey.Length;
+
+					var value = nv[key];
+
+					if (value.IsNullOrEmpty()) continue;
+
+					buffer[position++] = '=';
+					var escapedValue = Uri.EscapeDataString(value);
+					escapedValue.AsSpan().CopyTo(buffer.Slice(position));
+					position += escapedValue.Length;
+				}
+
+				return buffer.Slice(0, position).ToString();
+			}
+			finally
+			{
+				if (rentedFromPool is object)
+					ArrayPool<char>.Shared.Return(rentedFromPool, clearArray: false);
+			}
 		}
 
 		internal static void UpdateFromDictionary(this NameValueCollection queryString, Dictionary<string, object> queryStringUpdates,