This commit optimises the internal `ToQueryString` extension method on the `NameValueCollection` type. It avoids some intermediate string allocations by using a zero-alloc buffer to format the final query string.

- Adds a test for the extension method to validate existing behaviour is unaffected by the change of implementation.
- Adds the `System.Memory` package to support `Span<T>` usage across all target frameworks. This is registered conditionally for targets before NetStandard 2.1. I had some downgrade errors when using the latest package version, so I've aligned with the existing `System.Buffers` version for the time being.

I've opened this against master although it may be valuable to back port to existing supported versions?

The exact performance improvement varies based on the underlying collection, but in all cases resulted in less allocations and executes equal to or faster than the original code.

## .NET Core 2.1 Benchmark Results

```c#
private NameValueCollection _nvc = new NameValueCollection
{
    { "q", "title:\"The Right Way\" AND mod_date:[20020101 TO 20030101]" },
    { "from", "10000" },
    { "request_cache", bool.TrueString },
    { "size", "100" }
};
```

```
|   Method |     Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
| Original | 4.105 us | 0.0901 us | 0.2614 us |  1.00 |    0.00 | 0.3052 |     - |     - |    1288 B |
|      New | 3.652 us | 0.0730 us | 0.1898 us |  0.89 |    0.07 | 0.1755 |     - |     - |     736 B |
```

- Allocations reduced by 42.9%
- 11% faster

For smaller collections and key/value lengths, the allocation improvement can be greater. I've seen up to 77% on some test cases where no URI encoding need occur.

Closes #4951