@stevejgordon 2020-08-13
@Mpdreamz Modified the code slightly to assume worst case that every byte requires 3 char encoding. This avoids overflowing the buffer, even for strings entirely made up of emojis! Added suggested test + one other for 2 byte character.