@Mpdreamz 2020-08-13
The integration test failures are due to changes on the SNAPSHOT version of the server that we have not addressed yet. Opened https://github.com/elastic/elasticsearch-net/issues/4957 to fix those. 

The auto label failure is due to me creating a new label after this PR was opened and applying it to this PR.


Thanks for the updates @stevejgordon!